Personal_SaveApp/requirements.txt
docx2txt==0.8
pandas==1.5.3
pdfplumber==0.9.0
Pillow==9.4.0
Pillow==9.5.0
PyPDF2==3.0.1
streamlit==1.19.0
